+++ src/fh/component/mod.rs
@@ -1,0 +1,9 @@
+mod component;
+mod component_type;
+mod information;
+
+pub use {
+    component::Component,
+    component_type::ComponentType,
+    information::ComponentInformation,
+};